llm_sql/gradio/requirements.txt
langchain
openai
python-dotenv
huggingface_hub
pandas
gradio